feat: gate + harden user create/role-change abilities; exclude from wp.org build
Addresses WordPress.org plugin review feedback that custom user creation
routes can bypass security plugins (login throttling, password-policy
enforcers, etc.) that hook the native register/login flow.

Defence in depth — both at the same time:

1. **Hardening** of the two mutating user-management abilities:
   - Split `sd-ai-agent/create-user` and `sd-ai-agent/update-user-role`
     out of `UserAbilities.php` into a new `UserManagementAbilities.php`.
     The read-only `sd-ai-agent/list-users` ability stays in
     `UserAbilities.php` and ships in every build.
   - `create-user` permission_callback now requires BOTH `create_users`
     AND `promote_users` (previously only `create_users`).
   - `update-user-role` permission_callback requires BOTH `promote_users`
     AND `edit_users` (previously only `edit_users`).
   - Handlers re-verify caps inside the function body (defence-in-depth
     against direct PHP invocation that bypasses the REST permission
     layer) and return `ai_agent_forbidden` on failure.
   - Both handlers use `get_editable_roles()` for the role-escalation
     guard — the same gate `wp-admin/user-new.php` and
     `wp-admin/user-edit.php` use — so a non-administrator granted the
     ability cannot create or promote a user to administrator.
   - `update-user-role` refuses self-modification (prevents
     self-elevation) and refuses to modify a super-admin on multisite.

2. **Build-time removal** from the WordPress.org distribution:
   - New `SD_AI_AGENT_FEATURE_USER_MANAGEMENT` feature flag (defaults
     true) gates registration of both abilities at runtime via
     `Features::is_enabled()`.
   - `bin/build.sh --target=wporg` forces the constant to `false` in the
     bundled `superdav-ai-agent.php`, with a grep-friendly rationale
     comment for the WP.org review team.
   - `.distignore-wporg` physically strips
     `includes/Abilities/UserManagementAbilities.php` from the wp.org
     zip. The build validates the strip + `class_exists()` guard in
     `AbilitiesHandler` no-ops the registration call.

Verified the wp.org zip:
  - Contains `includes/Abilities/UserAbilities.php` (list-users only).
  - Does NOT contain `includes/Abilities/UserManagementAbilities.php`.
  - `superdav-ai-agent.php` line: `define( 'SD_AI_AGENT_FEATURE_USER_MANAGEMENT', false ); // wporg-build: custom user creation/role-change disabled per WP.org plugin review feedback ...`

Test changes:
  - Moved create/update tests from `UserAbilitiesTest.php` to new
    `UserManagementAbilitiesTest.php`; added 5 capability-gate /
    self-modification / cap-re-check tests.
  - Updated `SdAiAgentPublicFlagTest.php` and
    `AbilitySchemaIntegrityTest.php` to also register
    `UserManagementAbilities` so the registry-scan assertions cover the
    new ability class.
