+++ bin/build.sh
@@ -202,7 +202,7 @@ EXTRA
 	# prevents trivial bypass and gives the WP.org review team a single
 	# grep target to verify compliance.
 	if [ "$variant" = "wporg" ]; then
-		echo "==> [${variant}] Forcing plugin-builder + CLI + plugin-state + URL-install + file-write + scaffold-block-theme + wp-rest + wp-cli dispatcher + benchmark feature flags to false..."
+		echo "==> [${variant}] Forcing plugin-builder + CLI + plugin-state + URL-install + file-write + scaffold-block-theme + wp-rest + wp-cli dispatcher + benchmark + user-management feature flags to false..."
 		local main_file="${dest}/superdav-ai-agent.php"
 
 		# The feature flags this build target forces off are:
@@ -213,8 +213,9 @@ EXTRA
 		# SD_AI_AGENT_FEATURE_FILE_WRITE,
 		# SD_AI_AGENT_FEATURE_SCAFFOLD_BLOCK_THEME,
 		# SD_AI_AGENT_FEATURE_WP_REST_DISPATCHER,
-		# SD_AI_AGENT_FEATURE_WP_CLI_DISPATCHER, and
-		# SD_AI_AGENT_FEATURE_BENCHMARK. Each entry is paired with a
+		# SD_AI_AGENT_FEATURE_WP_CLI_DISPATCHER,
+		# SD_AI_AGENT_FEATURE_BENCHMARK, and
+		# SD_AI_AGENT_FEATURE_USER_MANAGEMENT. Each entry is paired with a
 		# short rationale that ends up as an inline comment in the bundled
 		# main plugin file (a grep target the WP.org review team can use).
 		local -a flags=(
@@ -227,6 +228,7 @@ EXTRA
 			"SD_AI_AGENT_FEATURE_WP_REST_DISPATCHER:low-level REST dispatcher disabled per WP.org Guideline 4"
 			"SD_AI_AGENT_FEATURE_WP_CLI_DISPATCHER:shell-exec wp-cli dispatcher disabled per WP.org Guideline 4"
 			"SD_AI_AGENT_FEATURE_BENCHMARK:wp-cli benchmark suite with arbitrary --log-dir disabled per WP.org reviewer feedback"
+			"SD_AI_AGENT_FEATURE_USER_MANAGEMENT:custom user creation\/role-change disabled per WP.org plugin review feedback - bypasses security plugins on native register\/login flow"
 		)
 
 		# Replace each `defined() || define( 'NAME', true )` line with a
@@ -266,6 +268,7 @@ EXTRA
 			"${dest}/includes/Abilities/WpCliAbilities.php"
 			"${dest}/includes/Benchmark"
 			"${dest}/includes/CLI/BenchmarkCommand.php"
+			"${dest}/includes/Abilities/UserManagementAbilities.php"
 		)
 		local p
 		for p in "${stripped_paths[@]}"; do
